refactor into small api helper
